+++ web/tests/acceptance/authenticated/my/documents-test.ts
@@ -4,7 +4,7 @@ import { module, test } from "qunit";
 import { authenticateSession } from "ember-simple-auth/test-support";
 import { MirageTestContext, setupMirage } from "ember-cli-mirage/test-support";
 import { getPageTitle } from "ember-page-title/test-support";
-import { TEST_USER_2_EMAIL, TEST_USER_EMAIL } from "hermes/mirage/mirage-utils";
+import { TEST_USER_2_EMAIL, TEST_USER_EMAIL } from "hermes/mirage/utils";
 
 const SORTABLE_HEADER = "[data-test-attribute=modifiedTime]";
 const OWNER_FILTER = "[data-test-owner-filter]";
