+++ web/tests/integration/components/document/sidebar-test.ts
@@ -9,7 +9,7 @@ import AuthenticatedUserService, {
 } from "hermes/services/authenticated-user";
 import { HermesDocument } from "hermes/types/document";
 import { HermesDocumentType } from "hermes/types/document-type";
-import { TEST_USER_2_EMAIL, TEST_USER_EMAIL } from "hermes/mirage/mirage-utils";
+import { TEST_USER_2_EMAIL, TEST_USER_EMAIL } from "hermes/mirage/utils";
 
 const SUMMARY_CONTAINER = "[data-test-document-summary]";
 const SUMMARY_EMPTY_STATE = `${SUMMARY_CONTAINER} .empty-state-text`;
