+++ web/tests/integration/components/person-test.ts
@@ -3,10 +3,7 @@ import { setupRenderingTest } from "ember-qunit";
 import { render } from "@ember/test-helpers";
 import { hbs } from "ember-cli-htmlbars";
 import { MirageTestContext, setupMirage } from "ember-cli-mirage/test-support";
-import {
-  TEST_USER_EMAIL,
-  authenticateTestUser,
-} from "hermes/mirage/mirage-utils";
+import { TEST_USER_EMAIL, authenticateTestUser } from "hermes/mirage/utils";
 
 const APPROVED_BADGE = "[data-test-person-approved-badge]";
