+++ web/tests/integration/components/project/jira-widget-test.ts
@@ -4,11 +4,8 @@ import { setupRenderingTest } from "ember-qunit";
 import { click, fillIn, find, render, waitFor } from "@ember/test-helpers";
 import { hbs } from "ember-cli-htmlbars";
 import { JiraIssue, JiraPickerResult } from "hermes/types/project";
-import {
-  TEST_JIRA_ISSUE_SUMMARY,
-  setWebConfig,
-} from "hermes/mirage/mirage-utils";
-import { TEST_JIRA_WORKSPACE_URL } from "hermes/mirage/mirage-utils";
+import { TEST_JIRA_ISSUE_SUMMARY, setWebConfig } from "hermes/mirage/utils";
+import { TEST_JIRA_WORKSPACE_URL } from "hermes/mirage/utils";
 
 const JIRA_ICON = "[data-test-jira-icon]";
 const ADD_JIRA_INPUT = "[data-test-add-jira-input]";
